Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
mistralai==0.4.2
google-generativeai==0.3.2
cohere==4.47
openai==1.12.0
requests==2.31.0
python-multipart==0.0.9
PyJWT==2.8.0
pydantic==2.6.1
typing-extensions==4.9.0
python-jose==3.3.0
